// wardn/frontend/lib/api/generated/model/runtimeLogEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.0.1
 */
import type { RuntimeLogEntryFields } from './runtimeLogEntryFields';
import type { RuntimeLogEntryLevel } from './runtimeLogEntryLevel';

export interface RuntimeLogEntry {
  fields: RuntimeLogEntryFields;
  id: string;
  level: RuntimeLogEntryLevel;
  message: string;
  timestamp: string;
}